Split out authorization db and checking into Authorizer
